+++ farm-fresh/src/App.js
@@ -1,11 +1,10 @@
 import React from 'react';
-
+import Consumer from './components/Consumer';
 
 function App() {
   return (
       <div className="App">
           <p>Farm Fresh</p>
-  
           <Consumer />
       </div>
   );
